run_util cmd, split out into root and run_util command in main.go, and report.html template
